requirements.txt
Django>=4.2.7
django-admin>=2.0.2
django-rest-swagger>=2.2.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
isoweek>=1.3.3
Jinja2>=3.1.2
MarkupSafe>=2.1.3
openapi-codec>=1.3.2
packaging>=23.2
Pillow>=10.1.0
python-dateutil>=2.8.2
requests>=2.31.0
TimeConvert>=3.0.13
python-dotenv>=1.0.0